moved the socket into a separated file
